# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for .",
        "title": "Rebellion Banking APII",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/banks/beeline/sims": {
            "get": {
                "description": "Returns all registered Beeline SIM cards.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beeline sims"
                ],
                "summary": "List Beeline SIMs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.SimResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Registers a Beeline SIM by 10-digit phone number without +7.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beeline sims"
                ],
                "summary": "Create Beeline SIM",
                "parameters": [
                    {
                        "description": "SIM payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.CreateSimRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.SimResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/beeline/sims/{number}": {
            "get": {
                "description": "Returns a Beeline SIM by phone number.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beeline sims"
                ],
                "summary": "Get Beeline SIM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "10-digit phone number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.SimResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a Beeline SIM and its payment history.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beeline sims"
                ],
                "summary": "Delete Beeline SIM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "10-digit phone number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/beeline/sims/{number}/config": {
            "get": {
                "description": "Returns config with effective balance after payment history for the SIM.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beeline config"
                ],
                "summary": "Get Beeline SIM config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "10-digit phone number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.ConfigResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/beeline/sims/{number}/config/balance": {
            "patch": {
                "description": "Updates initial balance for the SIM. Effective balance subtracts payment totals from history.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beeline config"
                ],
                "summary": "Update Beeline SIM base balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "10-digit phone number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Balance update payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.UpdateBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.ConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/beeline/sims/{number}/payments": {
            "get": {
                "description": "Returns payment history for the SIM.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beeline payments"
                ],
                "summary": "List Beeline SIM payments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "10-digit phone number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.PaymentResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a manual payment for the SIM. Use direction=outgoing for mobile commerce charge (requires receiverCard, min 924 RUB, 6.5%% commission). Use direction=incoming for balance refill (no card, no commission).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beeline payments"
                ],
                "summary": "Create Beeline SIM payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "10-digit phone number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.PaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_banks_beeline_transport_http.BeelineErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_modules_banks_beeline_transport_http.BeelineErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_modules_banks_beeline_transport_http.ConfigResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "baseBalance": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "paymentsTotal": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_modules_banks_beeline_transport_http.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "paidAt"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 13000
                },
                "direction": {
                    "type": "string",
                    "enum": [
                        "outgoing",
                        "incoming"
                    ],
                    "example": "outgoing"
                },
                "paidAt": {
                    "type": "string",
                    "example": "2026-05-23T12:07:47+03:00"
                },
                "receiverCard": {
                    "type": "string",
                    "example": "220094**0028"
                }
            }
        },
        "internal_modules_banks_beeline_transport_http.CreateSimRequest": {
            "type": "object",
            "required": [
                "number"
            ],
            "properties": {
                "number": {
                    "type": "string",
                    "example": "9680659702"
                }
            }
        },
        "internal_modules_banks_beeline_transport_http.PaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "commission": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "paidAt": {
                    "type": "string"
                },
                "receiverCard": {
                    "type": "string"
                },
                "simNumber": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_modules_banks_beeline_transport_http.SimResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_modules_banks_beeline_transport_http.UpdateBalanceRequest": {
            "type": "object",
            "required": [
                "balance"
            ],
            "properties": {
                "balance": {
                    "type": "number",
                    "example": 50000
                }
            }
        },
...
    }
}